fix: create utils.js and constants.js  and rename funcs
