+++ js/constants.js
@@ -1,0 +1,17 @@
+// Формы
+export const formEditProfile = document.querySelector('.popup__form');
+export const formAdd = document.querySelector('.popup__form_type_add');
+// Кнопки
+export const editBtn = document.querySelector('.button_type_edit');
+export const saveBtn = document.querySelector('.button_type_save');
+export const addBtn = document.querySelector('.profile__button-add');
+// Попапы
+export const popupEdit = document.querySelector('.popup_type_edit');
+export const popupAdd = document.querySelector('.popup_type_add');
+// инпуты
+export const profileName = document.querySelector('.profile__name');
+export const profileJob = document.querySelector('.profile__job');
+export const nameInput = formEditProfile.querySelector('.popup__form-input_type_name');
+export const jobInput = formEditProfile.querySelector('.popup__form-input_type_job');
+export const placeName = formAdd.querySelector('.popup__form-input_type_place-name'); // - название места инпут
+export const imageUrl = formAdd.querySelector('.popup__form-input_type_image-url'); //- ссылка на фото для инпута
