+++ js/formValidator.js
@@ -6,15 +6,6 @@ export const config = {
   inputErrorClass: 'popup__input-error_status_active',
   errorClass: 'popup__input-error'
 };
-export const profileName = document.querySelector('.profile__name');
-export const profileJob = document.querySelector('.profile__job');
-export const form = document.querySelector('.popup__form');
-export const nameInput = form.querySelector('.popup__form-input_type_name');
-export const jobInput = form.querySelector('.popup__form-input_type_job');
-export const formAdd = document.querySelector('.popup__form_type_add');
-export const placeName = formAdd.querySelector('.popup__form-input_type_place-name'); // - название места инпут
-export const imageUrl = formAdd.querySelector('.popup__form-input_type_image-url'); //- ссылка на фото для инпута
-
 export class FormValidator {
   constructor(config, formSelector){
     this._config = config;
